\name{stiagent-package}
\alias{stiagent-package}
\alias{stiagent}
\docType{package}
\title{
\packageTitle{stiagent}
}
\description{
\packageDescription{stiagent}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{stiagent}
\packageIndices{stiagent}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{stiagent}

Maintainer: \packageMaintainer{stiagent}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
